Move websocket internals into plumbing submodule
This will be the basis for core reuse between WebSocketClient for
different protocol versions.
